qvm_dashboard/requirements.txt
pandas
streamlit
pyyaml
plotly
pytest
